Remove unused headers and change test organization

+++ include/test/utils.h
@@ -1,13 +1,13 @@
 #ifndef ZORRO_TEST_UTILS_H
 #define ZORRO_TEST_UTILS_H
 
-#include "engine.h"
+#include <stdio.h>
 #include <stdlib.h>
 
 struct Lines;
 
 char *
-lines_nth(struct Lines *lines, long long i);
+lines_nth(struct Lines *lines, size_t i);
 
 size_t
 lines_count(struct Lines *lines);
